Set C# language version for all projects
Set C# language version to "5.0" for all C# modules. Previously this was enforced by ReSharper, however I recently discovered that this could be enforced by a built in feature of the .NET compiler which is more handy. I also set language level for the Android module by ReSharper as it was missing