Content Manager for plugins and code cleanup
I restructured plugin entry points (Campaign and UserInterface) to maintain their own Content Managers. The idea is that every plugin should should maintain their own assets, which will be especially useful when we need to load and unload plugins. Some code cleanup and organization was also required to implement this.