Fixed critical error
Commit fe39350e37403bc3e78d23b23a166217ad2ccd95 introduced a severe error of missing files that will block anyone from building the project. This commit should remedy that.

I also simplified the code referencing the directories to the assets, so that we could reference them simply by filename instead of entire directory path.

Finally I removed the gitignore for the files that are auto generated by EmptyKeys, as apparantly the generator program only works on Windows, we'll need the actually generated files for non Windows developers.