Modified ToolsVersion
Changed ToolsVersion into highest possible value that is supported on Linux builds. Most projects was set to 14, although highest version supported by Linux is 12. I also discovered that some projects had their ToolsVersion set to 4, which may be an artifact from when I have attempted to build on Linux in the past, and just set an arbitrary value.