Split SierraVga into 2 modules
I split the SierraVga module into 2 modules, 1 containging the C# code and one with the XAML designer forms. The problem I needed to solve was that the utility program for generating the C# code from the XAML forms works only on Windows which puts developers on other platforms at a disadvantage.