Upgraded every module .NET version (v4.5.1)
Upgraded the .NET version of every module to v.4.5.1, this is the most recent supported version by Mono in MonoDevelop on Linux. This upgrade includes the projects that orignally was at v2.0 because support for this version is getting increasingly tedious to manage on newer OS (Windows 10, Ubuntu 16). These projects were orignally set to v2.0 to maintain compability with as many frameworks and engines as possible, but it's just not worth the hassle anymore.