Added OS X build to Travis and FAKE build scripts
Like the title says I have adjusted the Travis config to include OS X and tweaked it so that it works on both Linux and OS X. During this ordeal I decied to add FAKE to the project, by using Common Language Interfacing I can gather all command line build scripts into a single place using batch and shell scripts only as native entry points.

FAKE is completely optional to use, it is helpful for low level CI servers like Travis which rely on scripts or terminal commands for building, but it can also be a useful tool for developers who want to work in text editors like Vim or Emacs instead of dedicated IDEs.