Verb interface
This introduces the UI component for the verb bar of the Sierra VGA template. I have added the top bar (mostly for show), the actual verb bar and some mouse cursors. Pressing any of the verb icons (walk, look, hand, talk) should change the mouse cursor. The rest of the icons prints placeholder messages to the console.