Reworking of Windows Installer
The WiX based Windows Installer has been replaced with Advanced Installer. Our project uses the Professional template so a license is required to work on the installer.

I've also added the new feature that prerequisite software is installed, specifically .NET v4.5.1 and DirectX v9.0c.

On top of that the installer UI now has a Dark Metro theme.