Crash Logs
Introduced a logging feature that will store some basic information about the game in case of a crash, and save it to a .txt file on the computer. On Windows the save location will be "My Documents/Hero6/logs".

I programmed it so that this will only occur with Release builds, not on Debug builds, as I want to still be able to catch exceptions with the debugger rather than read them from a .txt file on disk.