Reworked input
Input is now completely removed from the Hero6 startup modules, and is designed to be handled entirely from the UI modules.

I also removed the touch controls without providing alternatives, however I think it is okay to do so as it turns out the original code for touch input was no longer working.