Updated Readme with MonoGame v3.6
I updated the url so it pointed to the right version. I also removed a line "how to setup your IDE with MonoGame" as this implies MonoGame integrates into the IDE, which is false, it installs as an msbuild extension, which is why it works out of the box with Xamarin Studio and Rider after only one install.