Updated CI build scripts and cleaned up related legacy hack from .csproj
Updated Travis and Appveyor scripts so that they would use MonoGame v3.6 instead of v3.4. v3.4 was put here to circumvent a bug, with this I can also get rid of a legacy hack put in place in some of the .csproj files, where additional MonoGame build instructions that is only relevant for MonoGame pre v3.5 no longer has to exist.