Upgraded MonoGame from v3.5.1 to v3.6 via Paket
As the title says. I'm a little unhappy that this new release introduces more files via SDL2 that has to be installed on disk, making us even more reliant on making willing participants install additional software on their computers. However I've been sneaking through their bug tracker and I know that they're at least discussing and planning to better this situation.

In order to try this commit it is a pre-requisite that MonoGame v3.6 is installed.