Removed gone .dll from Windows Installer
SharpDX.RawInput.dll is suddenly not copied to the output directory on build. I can't exactly figure out why this is, but everything seems to be in working order so I'm going to assume that it is by design. It's easy enough to add back to the installer if we were to need that .dll in the future.