Treat Warnings as errors
I set the .csproj property "Treat warnings as errors" to true, for most open-source projects I've seen this is standard practice. The perk to this is obvious, avoid warnings sneaking in on pull requests and have them stacking up on us. It can make coding a little more tedious I guess but in my experience this is not a problem.