#10 Support for Localization
Added .resx files to our campaign. This should be satisfactory for supporting multiple languages. It's not quite clear since we have so little content at this point, but my intention is that each plugin will have to provide their own "Strings.resx", I think this is the best way to make sure we support various locales when we support extending the game from third parties.

I have added the source files for the strings to the "assets" folder because I feel that is where they belong, as we use folders to divide and organize our licensed material. It's reasonable to say descriptive texts, dialoguies, etc. should be put under the "assets" folder as the work should fall under our licensing for assets.