Removed unconventional unit test
I removed some unit tests that relied on GPU hardware to initialize. I should never have made these unit tests in the first place, but I fell for the temptation of having tests that checks if the game actually starts. In reality however we had to make ugly special cases for the main game test assembly so it wouldn't run tests on systems missing GPUs, and the tests was to expensive perfomance wise to be integrated into development patterns like TTD or any form where unit tests should be run often.