Simplify Update and Draw argument list
This fixes a small annoyance I've had for a long time with the Update and Draw function calls. The line with the function name and arguments is so long I end up breaking it over multiple lines, however by simplifying some of the argument names(removing time from gameTime argument name) I can put the Update and Draw function to a single line, which is much nicer.