Set rule SA1028 to warning, code must not containt trailing whitespaces
I don't remember what reasons we had for demoting this rule into a "Info" but I'm going to veto this and set it back to a Warning check. The Info is not satisfactory for this as it only triggers when the actual code file is open in the editor, and we do not want trailing white spaces.