Fixed UI mouse click triggering even if UI was invisible
Every UI component was in charge of handling their own mouse click input, however the problem with this was that child components would have no awareness if their parents were rendering. Obviously if a parent UI compoenent is inactive, then the children should inherit that property. The solution was to rewrite the implementation so that the mouse click was handled from top at User Interface, and from there navigate through all children