Fix FAKE script
Fixed some errors in the FAKE build script. Certain dependencies assumed that all the built files ended in a standard "bin" folder for each project, rather than a single custom chosen path which was the case with our FAKE script. This is now fixed so that it goes with the former option, rather than the latter, the reason this problem was introduced was a misunderstanding in how the FAKE MsBuild command worked, as the first string it takes as a forced input argument is the build dir, however as it turns out it is completely legal to pass a empty string and it will then defer to standard behavior