Modified Localization project so that it automatically adds translations
We are now adding translated files by wildcard, so anyone can just add a file "Strings.my-language.resx" and it will be automatically included in the IDE and compiler. I think this solution is great because anyone can fork our project, create a translation and not have to worry about merge conflicts at any point.