Fixed that logger unit tests would arbitrarily fail
Log files was stored by using date as ID down to the second, however the second was sometimes rouded of differently when stored as a string variable and when it was stored as a filename on disk.

I solved this issue by getting rid of the date identifier completely in favor of a incrementing integer for every time the game is started.