Stopped using 2 rulesets and just go with 1
It's simpler this way. Before I had 2 rulesets, on for projects I wanted to document and 1 for implementation projects that didn't need documentation. The idea was that certain projects could be reused as frameworks so they needed documentation. However we have to make a choice whether we want to develop a game or if we want to develop technology. All documentation written ended up being very lazily written anyway. Adding documentation from now on will be a judgement call